refactor: Disable generation of Angular "vendor" chunks QUIC-1585
This change makes the files produced by a "dev" build and a "prod" build the same, so that we can use the same bundler for dev mode and production releases.

By default, Angular produces 4 files in dev (main, runtime, polyfills, vendor) and 3 in prod (main, runtime, polyfills)

Since we bundle all the output files into a single `.js` file for distribution, there is no need for the extra code splitting anyway, and this simplifies our bundling to always pull in the same 3 files.