ENG-1411: Resolve issue with no microphone audio (#392)
While this shouldn't be necessary it appears during testing that the
enableMicrophone/enableCamera/autoConnect fields are not converted from
strings. This results in them failing to compare against boolean true
causing the microphone/camera enable on the scene to effectively be
false. This may be happening (in part) due to the way the attribute
values are passed through into the typescript code. This change is a
work around that ensures the underlying value is a boolean. While
there's likely a better solution this appears to resolve the issue for
the imminent release.