ci: Upgrade Cypress to latest and get all tests passing ENG-1842 (#402)
Cypress is a test runner that executes UI test inside a browser
environment, allowing it to act as a user, verify the visibility (or
not) of elements for a given state of the data, and also to potentially
detect changes in the UI appearance between commits.

Our implementation of Cypress is fairly basic, but it runs a small suite
of integration tests which check things like "mic button toggling works"
and "content cards render correctly" etc.

You can run Cypress locally in the repo by:
- `npm run build:example` to populate the dist folder
- `npm run serve` to serev the dist folder at localhost:5050
- `npm run cypress:open` to launch the Cypress GUI which allow you to
explore and run available test suites

Our tests essentially load up the example pages and then wait for the
widget to connect to a real project, then runs some tasks on the widget
like clicking buttons, sending prompts to the DP to elicit a content
cards response etc.

The Cypress tests run in a headless browser on GitHub. We test against
Chrome and Edge currently (not FF that I could see)

Screen recording attached so you don't have to actually run it locally
to get an idea of what it does:


https://github.com/user-attachments/assets/b8d27875-fe42-4c8c-b4a2-de07166f72ec