chore: Upgrade Storybook to v8 and Vite to v4 ENG-1925 (#403)
Upgrade Storybook to v8 (the most recent version finally, yay!) to
enable future addition of Storybook plugins that use the new v8 APIs.

Vite v4 is a dependency of Storybook v8, so Vite was upgraded at the
same time.

Verified that `npm start` command also still serves html-based examples.

`vite-plugin-html` does not have a v4, so is still at v3 - it does not
appear to follow the major version changes of Vite.